[HRS-14] Combine Edit and Add function and change get employee function
