+++ src/app/features/employees-management/add-edit-employees/add-edit-employees.component.scss
@@ -1,0 +1,20 @@
+.full-width {
+  width: 100%;
+  margin-bottom: 12px;
+}
+
+.dialog-actions {
+  display: flex;
+  justify-content: space-between;
+  margin-top: 20px;
+}
+
+.left {
+  display: flex;
+  align-items: center;
+}
+
+.right {
+  display: flex;
+  gap: 8px;
+}
