+++ src/app/features/employees-management/add-edit-employees/add-edit-employees.component.spec.ts
@@ -1,0 +1,22 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { AddEditEmployeesComponent } from './add-edit-employees.component';
+
+describe('AddEditEmployeesComponent', () => {
+  let component: AddEditEmployeesComponent;
+  let fixture: ComponentFixture<AddEditEmployeesComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [AddEditEmployeesComponent],
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(AddEditEmployeesComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
